Specified classes and methods engine should adhere to.
Populated an architecture of sorts.

SceneTree will load the XML and render itself recursively.
It will iterate through model components which will at XML load time be populated with the specified model vertices to ask them to render themselves.
Vertices will be piled into a giant array in the model component and be rendered in steps of 3 for triangles.
Model files should have number of vertices at the head of the file to initialize the array.