Altered Engine to allow catmull-rom curves for translations. Also extended rotate to allow time-based rotation.
XML is read for time and points for a given translate if said translate has children and an AnimationComponent is filled out for the corresponding GroupComponent.

Extended possible operations in the order vector to include an animation translate (ANT) and animated rotate (ANR).

Added a timestamp global variable to extract the global time since render start at start of frame draw. Will be used to compute curve times and looping.

Added the catmull-rom matrix as a static global variable as well.